Move components from elements to its own folder
This should mirror the structure in the docs.
