+++ packages/palette/src/components/EntityHeader/index.tsx
