+++ packages/palette/src/components/LabeledRange/index.tsx
