+++ packages/palette/src/components/StaticCountdownTimer/index.tsx
