+++ packages/palette/src/elements/index.tsx
@@ -7,17 +7,13 @@ export * from "./Button"
 export * from "./Checkbox"
 export * from "./Collapse"
 export * from "./Colors"
-export * from "./EntityHeader"
 export * from "./Flex"
 export * from "./Grid"
 export * from "./Image"
 export * from "./Join"
-export * from "./Menu"
 export * from "./Message"
 export * from "./Pagination"
-export * from "./LabeledRange"
 export * from "./Link"
-export * from "./PriceRange"
 export * from "./ProgressBar"
 export * from "./ProgressBarTimer"
 export * from "./Radio"
@@ -29,11 +25,9 @@ export * from "./Slider"
 export * from "./Spacer"
 export * from "./Spinner"
 export * from "./StackableBorderBox"
-export * from "./StaticCountdownTimer"
 export * from "./Stepper"
 export * from "./Tabs"
 export * from "./TextArea"
-export * from "./TimeRemaining"
 export * from "./Toggle"
 export * from "./Tooltip"
 export * from "./Typography"
