+++ packages/palette/src/index.tsx
@@ -1,3 +1,4 @@
+export * from "./components"
 export * from "./elements"
 export * from "./helpers"
 export * from "./svgs"
